refactor(sysAdmin): shorten detail URL from /sysAdmin/dashboard/[id] to /sysAdmin/[id]
### URL 階層の簡素化
/sysAdmin/dashboard/ 階層が重複していたのでフラット化:
- 旧: /sysAdmin/dashboard/[communityId] (L2 detail)
- 新: /sysAdmin/[communityId]

### Routing
- /sysAdmin                        → 統合コミュニティ一覧 (変更なし)
- /sysAdmin/[communityId]          → L2 detail (移動)
- /sysAdmin/dashboard              → /sysAdmin に redirect (既存維持)
- /sysAdmin/dashboard/[communityId] → /sysAdmin/[communityId] に redirect
  (旧ブックマーク保険として残置)
- /sysAdmin/create                 → 静的ルートなので [communityId] より優先、変更なし

### 内部リンクの追従
- CommunityRow onClick: /sysAdmin/dashboard/${id} → /sysAdmin/${id}
- E2E スペックを新 URL + 統合ビュー構成に合わせて更新
  (対応内容も Phase 2 harness 整備のタイミングに再リファクタ予定)

47 test pass、tsc/eslint clean。
