+++ src/app/sysAdmin/dashboard/[communityId]/page.tsx
@@ -1,20 +1,14 @@
-"use client";
-
-import { use } from "react";
-import { CommunityDashboardDetail } from "@/app/sysAdmin/features/communityDetail/components/CommunityDashboardDetail";
+import { redirect } from "next/navigation";
 
 type Props = {
   params: Promise<{ communityId: string }>;
 };
 
-// ヘッダーの title は CommunityDashboardDetail 側でデータ取得後に動的に
-// community 名をセットする (useHeaderConfig をクライアント層で呼ぶため)。
-export default function SysAdminCommunityDetailPage({ params }: Props) {
-  const { communityId } = use(params);
-
-  return (
-    <div className="mx-auto max-w-7xl p-4">
-      <CommunityDashboardDetail communityId={communityId} />
-    </div>
-  );
+/**
+ * 旧 `/sysAdmin/dashboard/[communityId]` を新 `/sysAdmin/[communityId]` に
+ * リダイレクト。既存ブックマークや外部リンクの保険として残置。
+ */
+export default async function SysAdminCommunityDetailRedirect({ params }: Props) {
+  const { communityId } = await params;
+  redirect(`/sysAdmin/${communityId}`);
 }
